requirements.txt
aiohttp==3.7.4.post0
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
braceexpand==0.1.7
cffi==1.14.6
chardet==4.0.0
click==8.0.1
discord.py @ git+https://github.com/Rapptz/discord.py@45d498c1b76deaf3b394d17ccf56112fa691d160
idna==3.2
import-expression==1.1.4
jishaku==2.3.0
multidict==5.1.0
pycparser==2.20
PyNaCl==1.4.0
six==1.16.0
typing-extensions==3.10.0.2
yarl==1.6.3
youtube-dl==2021.6.6
